Move printer to cmd/crank/printer
